requirements.txt
psycopg2==2.9.9
sonyflake-py==1.3.0
streamlit==1.32.2
python-dotenv==1.0.1